cancunf-kernel: kernel-headers: Fix build use __u64 instead of __uint128_t

+++ kernel-headers/arch/arm64/include/uapi/asm/sigcontext.h
@@ -74,7 +74,7 @@ struct fpsimd_context {
 	struct _aarch64_ctx head;
 	__u32 fpsr;
 	__u32 fpcr;
-	__uint128_t vregs[32];
+	__u64 vregs[32];
 };
 
 /*
